Add files via upload
The Popcorn viewer was not working for players (as opposed to GMs) due to the fact that there was no way for the application to know that an update had been triggered for players. 
It was also breaking when a combatant token was removed from the scene without removing them from the combat first; this has been fixed so that errors are no longer thrown.